vers5oct

+++ wikiPr/wikiPr/Controllers/EntitiesContext.cs
@@ -13,11 +13,12 @@ using System.Data.SqlClient;
 using System.Configuration;
 using wikiPr.Models.Views;
 using System.Globalization;
+using System.Data.Entity.Migrations;
 
 
 
 namespace wikiPr.Controllers {
-    
+
     public class EntitiesContext : DbContext {
         public DbSet<Article> ArticleEntities { get; set; }
         public DbSet<Utilisateur> UtilisateurEntities { get; set; }
@@ -32,18 +33,19 @@ namespace wikiPr.Controllers {
             modelBuilder.Entity<Article>().Property(t => t.ArticleId).HasColumnName("ArticleId").HasColumnType("INT");
             modelBuilder.Entity<Article>().HasKey(t => t.ArticleId);
             modelBuilder.Entity<Article>().Property(t => t.ArticleId).HasDatabaseGeneratedOption(DatabaseGeneratedOption.Identity).IsRequired();
-           
+
             modelBuilder.Entity<Article>().Property(t => t.accederTitre).HasColumnName("Titre");
             modelBuilder.Entity<Article>().Property(t => t.accederContenu).HasColumnName("Contenu");
             modelBuilder.Entity<Article>().Property(t => t.accederDateModification).HasColumnName("DateModification");
             modelBuilder.Entity<Article>().Property(t => t.accederRevision).HasColumnName("Revision");
             modelBuilder.Entity<Article>().Property(t => t.accederIdContributeur).HasColumnName("IdContributeur");
 
-            modelBuilder.Entity<Article>().Property(t => t.accederTitre).IsUnicode(true).HasMaxLength(50).IsRequired(); 
+            modelBuilder.Entity<Article>().Property(t => t.accederTitre).IsUnicode(true).HasMaxLength(50).IsRequired();
             modelBuilder.Entity<Article>().Property(t => t.accederContenu).IsUnicode(true).HasMaxLength(4000);
             modelBuilder.Entity<Article>().Property(p => p.accederDateModification).HasColumnType("datetime");
             modelBuilder.Entity<Article>().Property(p => p.accederRevision).HasColumnType("INT");
             modelBuilder.Entity<Article>().Property(p => p.accederIdContributeur).HasColumnType("INT");
+
 
             //////
 
@@ -69,16 +71,16 @@ namespace wikiPr.Controllers {
 
         }
 
-        //public class EntitiesInitialize : CreateDatabaseIfNotExists<EntitiesContext> {
-        //    protected override void Seed(EntitiesContext context) {
-        //        base.Seed(context);
-        //        context.ArticleEntities.Add(new Article("article1", "WIKI", new DateTime(2017 - 08 - 09))); //nouveau article
-        //        context.UtilisateurEntities.Add(new Utilisateur("Prenom1", "Nom1", "courriel1@a.a", "aaaaaa", "fr")); //nouvel utilisateur
-        //        context.SaveChanges();
-        //        var initializer = new EntitiesInitialize(); //nouvelle instance de la classe
-        //        initializer.InitializeDatabase(new EntitiesContext());
-        //    }
-        //}
+        public class EntitiesInitialize : CreateDatabaseIfNotExists<EntitiesContext> {
+            protected override void Seed(EntitiesContext context) {
+                base.Seed(context);
+                context.ArticleEntities.Add(new Article("article1", "WIKI", new DateTime(2017 - 08 - 09))); //nouveau article
+                context.UtilisateurEntities.Add(new Utilisateur("Prenom1", "Nom1", "courriel1@a.a", "aaaaaa", "fr")); //nouvel utilisateur
+                context.SaveChanges();
+                //var initializer = new EntitiesInitialize(); //nouvelle instance de la classe
+                //initializer.InitializeDatabase(new EntitiesContext());
+            }
+        }
 
         public Article Find(int id) {
             EntitiesContext context = new EntitiesContext();
@@ -97,7 +99,7 @@ namespace wikiPr.Controllers {
         public Utilisateur FindById(int id) {
             EntitiesContext context = new EntitiesContext();
             context.Database.Log = Console.Write;
-           Utilisateur u = context.UtilisateurEntities.Where(m => m.Id.Equals(id)).FirstOrDefault();
+            Utilisateur u = context.UtilisateurEntities.Where(m => m.Id.Equals(id)).FirstOrDefault();
             return u;
         }
 
@@ -153,8 +155,13 @@ namespace wikiPr.Controllers {
 
         public static void supprimer(Article a) {
             EntitiesContext context = new EntitiesContext();
-            context.ArticleEntities.Remove(a);
-            context.SaveChanges();
+            // context.ArticleEntities.Attach(a);
+            //context.Entry(a).State = EntityState.Deleted;
+            if (a != null) {
+                context.ArticleEntities.Remove(a);
+                context.SaveChanges();
+            }
+            else { }            
         }
 
         public void supprimer(Utilisateur u) {
